Added Ship of the Sea
Ship has mesh colliders on the body, both sets of stairs as well as the crows nest. I left colliders off the rest to make sure she could be customized as needed. Materials added and a customized logo for FusionXR was added for the ship sails but can obviously be swapped out for a generic material